refactor(report): promote template + golden case management to a subdomain
Move every template-shaped concern out of `report/data/repository/`,
the bare `templateSelector.ts`, and the bare `judgeService.ts` into
a new `report/template/` subdomain that mirrors how `report/feedback/`
and other parent/child pairs are organised across the codebase
(account/community/config, transaction/incentiveGrant, etc).

  src/application/domain/report/
    template/                        # ★ new subdomain
      data/
        repository.ts                # 8 methods (templates + golden cases)
        interface.ts                 # IReportTemplateRepository (split out)
        type.ts                      # PrismaReportTemplate / PrismaReportGoldenCase
                                     # + reportTemplateSelect / reportGoldenCaseSelect
                                     # (moved from data/type.ts)
      selector.ts                    # ← report/templateSelector.ts
      judgeService.ts                # ← report/judgeService.ts

The shared parent `data/type.ts` keeps the report-entity-only items
(`reportSelect`, `PrismaReport`, `CommunitySummaryCursor`) — the
template / golden-case Prisma select shapes belong with the rest of
the template lifecycle.

Implementation bodies are unchanged. The 7 caller files
(service.ts / usecase.ts / provider.ts / presenter/index.ts /
controller/dataloader.ts and 3 unit tests) flip their import paths
to the new locations; nothing else moves.
